Edit and delete
Created the edit and delete functions on the items posted. An empty text on every row will be displayed when there are no data to be shown from the database. Started to make the categories blades and controller so that when the user adds a new item to have a select of categories instead of a input field.