Administrator access
Created a new groups table and linked it`s ids to the user table as a foreign id so that now we can separate administrator privileged users from the normal members, before creating anything we should either seed or manually add n the database 2 groups, the first one will be the admin having id 1 and the other being the normal member one. Made changed so that the administrator can edit and delete both posts and categories made by other users and changed the update methods so that the user_id is not updated anymore because it was changing the owner of posts and categories when edited by the administrator.