Admin functions
Made so that the administrator can create users and have a different edit panel for all of them that can change the group that the normal user belongs to. Changed how group was declared in the user.php and user in the group.php because they were not correctly made to call anything else beside the id of groups elements.